Update r7i-4x-modified config
Because this set of tests is designed to run on r7i-8x, but mimic
r7i-4x. We need to only use up to 8pcores for server container.
In our tests, we found because of AMX, use one vcore per pcore could
give us better performance, we tend to use only 8vcores.